add header lower to be included into platforms and refactor test
