+++ src/CBLUserAgent.hh
@@ -1,0 +1,56 @@
+//
+// CBLuserAgent.hh
+//
+// Copyright (c) 2022 Couchbase, Inc All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+
+#pragma once
+#include "CBL_Edition.h"
+
+
+using string = std::string;
+using alloc_slice = fleece::alloc_slice;
+// “CouchbaseLite”/<version> “-” <build #> ” (Java; ” <Android API> “;” <device id> “) ” <build type> “, Commit/” (“unofficial@” <hostname> | <git commit>) ” Core/” <core version>
+        //CouchbaseLite/3.1.0-SNAPSHOT (Java; Android 11; Pixel 4a) EE/debug, Commit/unofficial@HQ-Rename0337 Core/3.1.0
+        //"CouchbaseLite/0.0.0 (:fc1900d2+)-built from  branch, commit fc1900d2+CHANGES on Dec  7 2022 19:40:34(C;Apple OSX; <platform>) <build type>, Commit/unofficial@HQ-Rename0337 Core/0.0.0 (:fc1900d2+)"
+    static string createUserAgentHeader(){
+        string os, header;
+        alloc_slice coreVersion = c4_getVersion();
+        alloc_slice coreBuild = c4_getBuildInfo();
+#if defined (__APPLE__) && defined (__MACH__)
+    #if TARGET_IPHONE_SIMULATOR == 1
+        os = "Apple iOS Simulator";
+    #elif TARGET_OS_IPHONE == 1
+        os = "Apple iOS Device";
+    #elif TARGET_OS_MAC == 1
+        os = "Apple OSX";
+    #endif
+#elif __ANDROID__
+        os = "Android" + string(__ANDROID_API__);
+#elif  _WIN64
+        os = "Microsoft Windows (64-bit)";
+#elif _WIN32
+        os = "Microsoft Windows (32-bit)";
+#elif __linux__
+        os = "Linux";
+#else 
+        os = "Unknown OS";
+#endif
+        
+        header = "CouchbaseLite/" + string(CBLITE_VERSION) + "-" + string(CBLITE_BUILD_NUMBER) + " (" + os + ") Core/" + coreVersion.asString();
+        // header = "CouchbaseLite/" + "-" + "(C; " + os  + "; <platform>) <build type>, Commit/unofficial@HQ-Rename0337 Core/" + coreVersion.asString() + coreBuild.asString();
+        return header;
+}
