+++ src/exports/CBL_Exports.txt
@@ -234,3 +234,5 @@ CBLError_SetCaptureBacktraces
 
 CBLLog_BeginExpectingExceptions
 CBLLog_EndExpectingExceptions
+
+CBLReplicator_UserAgent
