+++ test/ReplicatorTest.cc
@@ -235,9 +235,10 @@ TEST_CASE_METHOD(ReplicatorTest, "Check userAgent header", "[Replicator]") {
     auto repl1 = CBLReplicator_Create(&config, &error);
     REQUIRE(repl1);
 
-    auto ccConfig = CBLReplicator_Config(repl1);
-    Dict headers = Dict(ccConfig->headers);
-    string userAgent = headers["userAgent"].asstring();
+    auto userAgent = slice(CBLReplicator_UserAgent(repl1)).asString();
+    // auto ccConfig = CBLReplicator_Config(repl1);
+    // Dict headers = Dict(ccConfig->headers);
+    // string userAgent = headers["userAgent"].asstring();
     CHECK_THAT(userAgent, Catch::StartsWith("CouchbaseLite/"));
     CBLReplicator_Release(repl1);
 }
